refactor: nft details
