--- components/NFTDetailsAccordion/NFTDeitalsAccordion.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './NFTDeitalsAccordion.module.scss';
-import { INFTDetailsAccordionProps } from './NFTDeitalsAccordion.types';
-const NFTDeitalsAccordion = ({}: INFTDetailsAccordionProps) => {
-  return <div className={styles.container}>NFTDeitalsAccordion</div>;
-};
-
-export default NFTDeitalsAccordion;
