--- components/NFTDetailsAccordion/NFTDeitalsAccordion.types.d.ts
@@ -1,3 +1,0 @@
-export interface INFTDetailsAccordionProps{
-  
-}
